+++ sofa-ark-parent/core-impl/archive/pom.xml
@@ -32,6 +32,7 @@
         <dependency>
             <groupId>junit</groupId>
             <artifactId>junit</artifactId>
+            <scope>test</scope>
         </dependency>
 
     </dependencies>
