+++ sofa-ark-parent/support/ark-plugin-maven-plugin/pom.xml
@@ -78,6 +78,7 @@
         <dependency>
             <groupId>junit</groupId>
             <artifactId>junit</artifactId>
+            <scope>test</scope>
         </dependency>
     </dependencies>
