+++ sofa-ark-parent/support/ark-springboot-starter/src/main/resources/META-INF/spring.factories
@@ -1,2 +1,2 @@
 org.springframework.context.ApplicationListener=\
-com.alipay.sofa.ark.springboot.ArkApplicationStartListener
+com.alipay.sofa.ark.springboot.listener.ArkApplicationStartListener
