+++ sofa-ark-parent/support/ark-springboot-starter/src/test/java/com/alipay/sofa/ark/ArkJUnit4RunnerTest.java
@@ -14,11 +14,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.alipay.sofa.ark.ide;
+package com.alipay.sofa.ark;
 
 import com.alipay.sofa.ark.bootstrap.ContainerClassLoader;
 import com.alipay.sofa.ark.container.tester.TestClassLoader;
-import com.alipay.sofa.ark.ide.runner.TestJUnit4Runner;
+import com.alipay.sofa.ark.runner.TestJUnit4Runner;
 import org.junit.*;
 import org.junit.runner.RunWith;
