+++ sofa-ark-parent/support/ark-springboot-starter/src/test/java/com/alipay/sofa/ark/springboot/SpringApplication.java
@@ -14,12 +14,19 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.alipay.sofa.ark.ide.springboot.facade;
+package com.alipay.sofa.ark.springboot;
+
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.ImportResource;
 
 /**
  * @author qilong.zql
  * @since 0.1.0
  */
-public interface SampleService {
-    String say();
+@ImportResource({ "classpath*:META-INF/sofa-ark-test/*.xml" })
+@SpringBootApplication
+public class SpringApplication {
+    public static void main(String[] args) {
+
+    }
 }
