+++ sofa-ark-parent/support/ark-springboot-starter/src/test/resources/META-INF/sofa-ark-test/sofa-ark-test.xml
@@ -3,6 +3,6 @@
        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
        xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans.xsd">
 
-    <bean id="sampleService" class="com.alipay.sofa.ark.ide.springboot.impl.SampleServiceImpl"/>
+    <bean id="sampleService" class="com.alipay.sofa.ark.springboot.impl.SampleServiceImpl"/>
 
 </beans>
