+++ sofa-ark-parent/support/ark-springboot-starter/src/test/resources/prepareForTest.sh
